+++ CHANGELOG.rst
@@ -1,6 +1,18 @@
 Changelog for Pulse3D
 =====================
 
+0.30.3 (2023-02-27)
+-------------------
+
+Added:
+^^^^^^
+- ``well_groups`` param added to ``PlateRecording`` to override the well groups in the H5 files
+- Added ``Platemap Group Metrics`` to the  ``aggregate-metrics`` sheet after individual well metrics
+
+Changed:
+^^^^^^^^
+- Removed ``PlateRecording.write_time_force_csv``
+
 0.30.2 (2023-02-22)
 -------------------